% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.ndcg.r
\name{amm.ndcg_at_k}
\alias{amm.ndcg_at_k}
\title{NDCG internal function}
\usage{
amm.ndcg_at_k(r, k = min(5, length(r)))
}
\arguments{
\item{r}{...}

\item{k}{...}
}
\value{
...
}
\description{
NDCG internal function. Source: https://www.kaggle.com/datadave/airbnb-recruiting-new-user-bookings/ndcg-score-r/code
}
\seealso{
...
}

