% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.mae.log.r
\name{amm_mae.log}
\alias{amm_mae.log}
\title{MAE log Metric for XGBoost}
\usage{
amm_mae.log(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
MAE log Metric for XGBoost
}
\seealso{
...
}

