% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.mae.r
\name{amm_mae}
\alias{amm_mae}
\title{MAE Metric for XGBoost}
\usage{
amm_mae(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
MAE Metric for XGBoost
}
\seealso{
...
}

