% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.mclogloss.r
\name{amm_mclogloss}
\alias{amm_mclogloss}
\title{MCLogLoss Metric for XGBoost}
\usage{
amm_mclogloss(actual, predicted, eps = 1e-15)
}
\arguments{
\item{actual}{vector}

\item{predicted}{vector}

\item{eps}{param}
}
\value{
Vector
}
\description{
MCLogLoss Metric for XGBoost. Be careful! May contain errors
}
\seealso{
...
}

