% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.ndcg.r
\name{amm_ndcg}
\alias{amm_ndcg}
\title{NDCG Metric for XGBoost}
\usage{
amm_ndcg(preds, dtrain)
}
\arguments{
\item{preds}{matrix or data.frame: one row for each observation, one column for each prediction. Columns are sorted from left to right descending in order of likelihood.}

\item{dtrain}{vector}
}
\value{
Vector: one row for each observation
}
\description{
NDCG Metric for XGBoost. Source: https://www.kaggle.com/datadave/airbnb-recruiting-new-user-bookings/ndcg-score-r/code
}
\seealso{
...
}

