% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.rmpse.r
\name{amm_rmpse}
\alias{amm_rmpse}
\title{RMPSE Metric for XGBoost}
\usage{
amm_rmpse(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
RMPSE Metric for XGBoost
}
\seealso{
...
}

