% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amm.rmpse.log.r
\name{amm_rmpse_log}
\alias{amm_rmpse_log}
\title{RMPSE log Metric for XGBoost}
\usage{
amm_rmpse_log(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
RMPSE log Metric for XGBoost
}
\seealso{
...
}

