% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo.caushyobj.log.r
\name{amo.caushyobj.log}
\alias{amo.caushyobj.log}
\title{Cauchy Log objective for XGBoost}
\usage{
amo.caushyobj.log(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
Cauchy Log objective for XGBoost. Be careful! May contain errors. Source: https://www.kaggle.com/dmi3kno/allstate-claims-severity/farons-xgb-starter-with-custom-objective
}
\seealso{
...
}

