% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo.caushyobj.r
\name{amo.caushyobj3}
\alias{amo.caushyobj3}
\title{Cauchy objective 3 for XGBoost}
\usage{
amo.caushyobj3(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
Cauchy objective 3 for XGBoost. Source: https://www.kaggle.com/dmi3kno/allstate-claims-severity/farons-xgb-starter-with-custom-objective
}
\seealso{
...
}

