% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo.fairobj.r
\name{amo.fairobj1.5}
\alias{amo.fairobj1.5}
\title{Fair objective 1.5 for XGBoost}
\usage{
amo.fairobj1.5(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
Fair objective 1.5 for XGBoost. Source: https://www.kaggle.com/dmi3kno/allstate-claims-severity/farons-xgb-starter-with-custom-objective
}
\seealso{
...
}

