% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo.fairobj.r
\name{amo.fairobje}
\alias{amo.fairobje}
\title{Fair objective e for XGBoost}
\usage{
amo.fairobje(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
Fair objective e for XGBoost. Source: https://www.kaggle.com/dmi3kno/allstate-claims-severity/farons-xgb-starter-with-custom-objective
}
\seealso{
...
}

