% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amo.logregobj.r
\name{amo.logregobj}
\alias{amo.logregobj}
\title{Logreg objective for XGBoost}
\usage{
amo.logregobj(preds, dtrain)
}
\arguments{
\item{preds}{vector}

\item{dtrain}{vector}
}
\value{
Vector
}
\description{
Logreg objective for XGBoost. Source: https://github.com/dmlc/xgboost/blob/master/R-package/demo/custom_objective.R#L14
}
\seealso{
...
}

